#2199 Moved config.go to its own subpackage.
