+++ internal/support/logging/config/config.go
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  *******************************************************************************/
-package logging
+package config
 
 import (
 	"github.com/edgexfoundry/edgex-go/internal/pkg/bootstrap/interfaces"
